moved generate stat to the right place
