+++ src/components/PathwayScoresView.js
@@ -3,11 +3,12 @@ import PureComponent from './PureComponent';
 import PropTypes from 'prop-types';
 import CanvasDrawing from "../CanvasDrawing";
 import DrawFunctions from '../functions/DrawFunctions';
-import {partition, sum, sumInstances} from '../functions/util';
+import {partition, sumInstances} from '../functions/util';
 import spinner from './ajax-loader.gif';
-import {pick, pluck, flatten} from 'underscore';
 import SVGLabels from "./SVGLabels";
 import {hierarchicalSort,clusterSort} from '../functions/SortFunctions';
+import {getGenePathwayLookup, getCopyNumberValue, getMutationScore} from '../functions/DataFunctions';
+import {pick, pluck, flatten,sum,range, times} from 'underscore';
 
 
 const REFERENCE_LABEL_HEIGHT = 150;
@@ -207,6 +208,87 @@ function pruneColumns(data, pathways, min) {
     };
 }
 
+/**
+ * For each expression result, for each gene listed, for each column represented in the pathways, populate the appropriate samples
+ *
+ * @param expression
+ * @param copyNumber
+ * @param geneList
+ * @param pathways
+ * @param samples
+ * @param filter
+ * @param min
+ * @param key
+ * @returns {any[]}
+ */
+   function associateData(expression, copyNumber, geneList, pathways, samples, filter, min, key){
+    filter = filter.indexOf('All') === 0 ? '' : filter;
+    let returnArray = times(pathways.length, () => times(samples.length, () => 0))
+    let sampleIndex = new Map(samples.map((v, i) => [v, i]));
+    let genePathwayLookup = getGenePathwayLookup(pathways);
+
+    // TODO: we should lookup the pathways and THEN the data, as opposed to looking up and then filtering
+    if (!filter || filter === 'Mutation') {
+        for (let row of expression.rows) {
+            let effectValue = getMutationScore(row.effect, min);
+            let pathwayIndices = genePathwayLookup(row.gene);
+
+            for (let index of pathwayIndices) {
+                returnArray[index][sampleIndex.get(row.sample)] += effectValue;
+            }
+        }
+    }
+
+
+    if (!filter || filter === 'Copy Number') {
+
+        // get list of genes in identified pathways
+        for (let gene of geneList) {
+            // if we have not processed that gene before, then process
+            let geneIndex = geneList.indexOf(gene);
+
+            let pathwayIndices = genePathwayLookup(gene);
+            let sampleEntries = copyNumber[geneIndex]; // set of samples for this gene
+            // we retrieve proper indices from the pathway to put back in the right place
+
+            // get pathways this gene is involved in
+            for (let index of pathwayIndices) {
+                // process all samples
+                for (let sampleEntryIndex in sampleEntries) {
+                    let returnValue = getCopyNumberValue(sampleEntries[sampleEntryIndex]);
+                    if (returnValue > 0) {
+                        returnArray[index][sampleEntryIndex] += returnValue;
+                    }
+                }
+            }
+        }
+
+    }
+
+    // let scoredColumn = returnArray.map(a => sum(a));
+    //
+    // let keyString = key + '';
+    //
+    // console.log('top-level keyString: ' + keyString);
+    //
+    // // TODO: add associated data to that app
+    // let appData = JSON.parse(JSON.stringify(this.state.apps));
+    // appData[key].associatedData = returnArray;gg
+    // appData[key].pathways = pathways;
+    // appData[key].scoredColumn = scoredColumn;
+
+    // console.log('appData');
+    // console.log(appData);
+    // let statBox = MultiXenaGoApp.generateStats(appData);
+    //
+    // this.setState({
+    //     apps: appData,
+    //     statBox: statBox,
+    // });
+
+    return returnArray;
+
+};
 
 
 
@@ -220,8 +302,8 @@ export default class PathwayScoresViewCache extends PureComponent {
     render() {
         let {cohortIndex,dataMunger,selectedPathways, selectedSort, min, filter, geneList, filterPercentage, data: {expression, pathways, samples, copyNumber, referencePathways}} = this.props;
         console.log('data munger');
-        
-        let associatedData = dataMunger(expression, copyNumber, geneList, pathways, samples, filter, min,cohortIndex);
+
+        let associatedData = associateData(expression, copyNumber, geneList, pathways, samples, filter, min,cohortIndex);
         let filterMin = Math.trunc(filterPercentage * samples.length);
 
         let prunedColumns = pruneColumns(associatedData, pathways, filterMin);
@@ -238,6 +320,10 @@ export default class PathwayScoresViewCache extends PureComponent {
                 returnedValue = clusterSort(prunedColumns);
                 break;
         }
+
+        dataMunger(returnedValue);
+
+
 
         if (referencePathways) {
             let referenceWidth = Math.max(minWidth, minColWidth * referencePathways.length);
